requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yt-dlp>=2025.1.0
musicbrainzngs>=0.7.1
slowapi>=0.1.9
python-dotenv>=1.0.0
pydantic>=2.0
httpx>=0.27.0
supabase>=2.4.0
